Move email and sms test to own files
